+++ api-module/src/main/java/org/c4marathon/assignment/infra/config/SwaggerConfiguration.java
