+++ api-module/src/main/java/org/c4marathon/assignment/member/dto/MemberRegistrationRequestDto.java
@@ -1,0 +1,4 @@
+package org.c4marathon.assignment.member.dto;
+
+public record MemberRegistrationRequestDto(String name, String email, String password) {
+}
