+++ api-module/src/main/java/org/c4marathon/assignment/settlement/SettlementController.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.controller;
+package org.c4marathon.assignment.settlement;
 
 import java.util.List;
