+++ api-module/src/main/java/org/c4marathon/assignment/transfer/dto/TransferRequestDto.java
@@ -1,6 +1,6 @@
-package org.c4marathon.assignment.dto.transfer;
+package org.c4marathon.assignment.transfer.dto;
 
-import org.c4marathon.assignment.common.validation.DifferentAccounts;
+import org.c4marathon.assignment.transfer.validation.DifferentAccounts;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
