+++ api-module/src/main/java/org/c4marathon/assignment/transferlog/dto/TransferLogCursorPageResponseDto.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.dto.transferlog;
+package org.c4marathon.assignment.transferlog.dto;
 
 import java.time.LocalDateTime;
 import java.util.List;
